+++ lib/constants/constants.dart
@@ -1,0 +1,3 @@
+class Constants {
+  static const String commonLogsDirectory = 'swift_log_logs';
+}