+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Capturing exceptions with customizable tags, subtags, and log messages
+* Capturing events with event names, log messages, and log levels
+* Saving events to device storage (They are cleared out as soon as they get pushed to your logs API)